requirements.txt
streamlit>=1.32.0
anthropic>=0.21.0
openai>=1.14.0
pydantic>=2.6.4
pymupdf>=1.24.0      # preferred: vision-based PDF extraction (images + text)
PyPDF2>=3.0.1        # fallback: text-only extraction if pymupdf unavailable
google-api-python-client>=2.122.0
google-auth>=2.28.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
